修复: 设备 ziggurat 尾部无符号负号回绕 (-(u64>>11) 模 2^64 -> log1p(2048) 永不接受) -- 一个 bug 三重化身: 无界形态挂起(='NVRTC嫌疑'真身)/限次形态自旋(=RNG 385ms 病理)/耗尽返回未剥离值; 2 文件 4 处 + e0o 探针, 逐位等价修复(cast 先于取负); tail 门逐位=opus, drift_orn 385->3.0ms, E0p 积分行 384->2.30ms (167x), 分量总 1862 ns/day; P3 翻转 2->0 (0.049% 归因证伪回收); E0o/forallK/E0n 全绿

+++ gpu/integrate.py
@@ -108,10 +108,15 @@ extern "C" __global__ void drift_orn(
                 if (rabs < ki[idx]) break;
                 if (idx == 0) {
                     for (int tail = 0; tail < 100000; ++tail) {
+                        // cast BEFORE negation (see rand.py
+                        // ziggurat_normal note -- the u64 wrap made
+                        // this loop spin to exhaustion, returning the
+                        // unstripped value, and made the unbounded
+                        // forms hang)
                         double xx = -nor_inv_r
-                            * log1p(-(rng.next_u64() >> 11)
+                            * log1p(-((double)(rng.next_u64() >> 11))
                                     * (1.0 / 9007199254740992.0));
-                        double yy = -log1p(-(rng.next_u64() >> 11)
+                        double yy = -log1p(-((double)(rng.next_u64() >> 11))
                                            * (1.0 / 9007199254740992.0));
                         if (yy + yy > xx * xx) {
                             g1 = ((rabs >> 8) & 0x1) ? -(nor_r + xx)
